update : laporan penjualan

+++ database/migrations/2023_09_23_034415_create_detail_transaksi_table.php
@@ -17,9 +17,9 @@ class CreateDetailTransaksiTable extends Migration
             $table->id();
             $table->integer('barang_id');
             $table->integer('transaksi_id');
-            $table->decimal('harga', 10, 2);
+            $table->integer('harga');
             $table->integer('jumlah');
-            $table->decimal('total', 10, 2);
+            $table->integer('total');
             $table->timestamps();
         });
     }
